requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pytesseract==0.3.10
pillow==10.4.0
httpx==0.27.0
